// src/_services/api/schema/center/center.zod.ts
export const criarConfiguracaoImpressaoBodyDataMaximaPercentualMin = -2147483648;
export const criarConfiguracaoImpressaoBodyDataMaximaPercentualMax = 2147483647;

export const criarConfiguracaoImpressaoBody = zod.object({
  "id": zod.string(),
  "tipoImpressao": zod.enum(['TRANSPORTE', 'CLIENTE']),
  "quebraPalete": zod.boolean(),
  "tipoQuebra": zod.union([zod.enum(['LINHAS', 'PERCENTUAL']),zod.null()]),
  "valorQuebra": zod.union([zod.string(),zod.null()]),
  "separarPaleteFull": zod.boolean(),
  "separarUnidades": zod.boolean(),
  "exibirInfoCabecalho": zod.union([zod.enum(['PRIMEIRO', 'TODOS', 'NENHUM']),zod.null()]),
  "segregarFifo": zod.union([zod.array(zod.string()),zod.null()]),
  "dataMaximaPercentual": zod.number().min(criarConfiguracaoImpressaoBodyDataMaximaPercentualMin).max(criarConfiguracaoImpressaoBodyDataMaximaPercentualMax),
  "createdAt": zod.string(),
  "updatedAt": zod.string(),
  "centerId": zod.string(),
  "atribuidoPorId": zod.union([zod.string(),zod.null()]),
  "empresa": zod.string(),
  "tipoImpressaoConferencia": zod.enum(['TRANSPORTE', 'CLIENTE']),
  "ordemConferencia": zod.union([zod.array(zod.string()),zod.null()]),
  "ordemFifo": zod.union([zod.array(zod.string()),zod.null()]),
  "ordemPaletes": zod.union([zod.array(zod.string()),zod.null()]),
  "ordemPicking": zod.union([zod.array(zod.string()),zod.null()]),
  "ordemUnidades": zod.union([zod.array(zod.string()),zod.null()])
})